Sort I18n.available_locales (#405)
We ran into an unexpected issue that different machines return `I18n.available_locales` in different order. Did not find out the root cause.

The impact was the digest used in the view is different from the one in the filename.

In case anyone runs into the same issue, this should fix it.