Improve Brazilian Portuguese translations
This commit aims to improve the Brazilian Portuguese translations of
trestle, as in its current state, it shows gramatically incorrect
messages such as 'Novo Mensagem' and 'O Loja foi deletado com sucesso'.

Thus, it changes the translations to something more neutral, in order
to ensure that everything is properly written.